+++ src/FileJobClient.ts
@@ -7,6 +7,7 @@ import { JobClient } from "./JobClient";
 import { ApiError } from "./ApiError";
 import { Logger } from "./Logger";
 import { humanReadToBytes, byteArrayToChunks, fileToChunks } from "./nodeUtils";
+import { DEFAULT_URL } from "./constants";
 
 import type {
   ClassInitiator,
@@ -20,24 +21,20 @@ interface SubmitOptions extends http.RequestOptions {
 
 export class FileJobClient {
   logger: Logger;
+  jobClient: JobClient;
   readonly baseUrl: string;
   readonly headers: {
     Authorization: string;
   };
-  jobClient: JobClient;
 
   /**
    * Creates a JobClient
    * @param {Object} config object
-   * @param {string} config.url - base url of modzy api (i.e.: https://app.modzy.com/api)
+   * @param {string} config.url - base url of modzy api (i.e.: https://app.modzy.com)
    * @param {string} config.apiKey - user's API key
    */
-  constructor({
-    url = "https://app.modzy.com/api",
-    apiKey,
-    logging,
-  }: ClassInitiator) {
-    this.baseUrl = url.endsWith("/") ? url.substring(0, url.length - 1) : url;
+  constructor({ url = DEFAULT_URL, apiKey, logging }: ClassInitiator) {
+    this.baseUrl = url;
     this.headers = {
       Authorization: `ApiKey ${apiKey}`,
     };
@@ -46,7 +43,7 @@ export class FileJobClient {
   }
 
   _getFeatures() {
-    const requestURL = `${this.baseUrl}/jobs/features`;
+    const requestURL = `${this.baseUrl}/api/jobs/features`;
     this.logger.debug(`_getFeatures GET ${requestURL}`);
 
     return axios
@@ -113,7 +110,7 @@ export class FileJobClient {
     inputValue: any,
     chunkSize: number
   ) {
-    const requestURL = `${this.baseUrl}/jobs/${job.jobIdentifier}/${inputItemKey}/${dataItemKey}`;
+    const requestURL = `${this.baseUrl}/api/jobs/${job.jobIdentifier}/${inputItemKey}/${dataItemKey}`;
     this.logger.debug("_appendInput called");
     let iterator;